refactor(controller): refresh collateral risk params on supply/withdraw
Collateral LTV, liquidation bonus, and (HF-gated) liquidation threshold now
sync from the effective asset config on every user supply/withdraw via the new
helpers/risk_params module, so positions reflect current governance config
without waiting for keeper propagation. Loosening applies immediately;
tightening requires post-update HF >= 1.05.

Oracle:
- rename OraclePolicy::allows_missing_twap_fallback -> allows_degraded_dual_source
- migrate tolerance constants and calculate_tolerance_range from i128 to u32
- move validate_oracle_bounds into tolerance::require_last_tolerance_gt_first
- replace PriceComponents/resolve_price with ResolvedOracleComponents::to_abi_prices
- move reads_same_feed_as onto OracleSourceConfigInput
- normalize feed timestamp checks via validate_positive_price_timestamps

Pool: pre-check flash-loan repay allowance so under-approval surfaces as
InvalidFlashloanRepay instead of a raw token error.

Config: replace MAX_SUPPLY/BORROW_POSITIONS consts with runtime PositionLimits;
drop dead consts, events, and type aliases.

Verification: restructure test-harness src and tests into domain subdirectories;
add Certora confs/specs (oracle compose, tolerance math, solvency, boundary) and
hosted-prover scripts; route Certora summaries via in-source cfg instead of a
module #[path] override. CI runs Certora compile + rule coverage; Makefile gains
certora-wasm / wasm-artifacts / certora targets.

+++ common/src/constants/oracle.rs
@@ -1,11 +1,14 @@
 // Protocol-wide upper bound for operator-supplied per-asset sanity caps.
 pub const MAX_REASONABLE_PRICE_WAD: i128 = 1_000_000_000 * super::WAD;
 
-pub const MIN_FIRST_TOLERANCE: i128 = 50;
+/// Minimum first-price tolerance input (BPS) for `configure_market_oracle`.
+pub const MIN_FIRST_TOLERANCE: u32 = 50;
 
-pub const MAX_FIRST_TOLERANCE: i128 = 5_000;
+/// Maximum first-price tolerance input (BPS).
+pub const MAX_FIRST_TOLERANCE: u32 = 5_000;
 
-pub const MIN_LAST_TOLERANCE: i128 = 150;
+/// Minimum last-price tolerance input (BPS).
+pub const MIN_LAST_TOLERANCE: u32 = 150;
 
-/// Maximum last-price tolerance in BPS.
-pub const MAX_LAST_TOLERANCE: i128 = 5_000;
+/// Maximum last-price tolerance input (BPS).
+pub const MAX_LAST_TOLERANCE: u32 = 5_000;
